Dependency Injection is go

+++ PaperPlane/app/src/main/java/com/ambergleam/android/paperplane/util/DistanceUtils.java
@@ -6,7 +6,7 @@ public class DistanceUtils {
 
     private static final String DECIMAL_FORMAT_DISTANCE = "#,###,###";
 
-    public static String formatDistance(long distance_ft) {
+    public static String formatDistance(int distance_ft) {
         DecimalFormat formatter = new DecimalFormat(DECIMAL_FORMAT_DISTANCE);
         String formattedDistance = formatter.format(distance_ft).concat(" ft");
         return formattedDistance;
